requirements.txt
python-telegram-bot[socks]==20.0